git-svn-id: https://svn.code.sf.net/p/numere/source/trunk@1302 d50cd26f-24d5-421e-9e7d-30b8b4216d91

+++ kernel/core/datamanagement/tablecolumnimpl.hpp
@@ -366,7 +366,7 @@ class CategoricalColumn : public TableColumn
     private:
         enum {CATEGORICAL_NAN = -1};
 
-        std::vector<size_t> m_data;
+        std::vector<int> m_data;
         std::vector<std::string> m_categories;
 
     public:
